Refactor out utility routines into separate module
