+++ src/fn/upload-custom-translations.js
@@ -1,4 +1,5 @@
 const fs = require('../lib/sync-fs');
+const utils = require('../lib/translation-file-utils');
 const skipFn = require('../lib/skip-fn');
 const warn = require('../lib/log').warn;
 const pouch = require('../lib/db');
@@ -18,52 +19,16 @@ module.exports = (projectDir, couchUrl) => {
       return Promise.all(fs.readdir(dir)
         .filter(name => FILE_MATCHER.test(name))
         .map(fileName => {
-          var translations = propertiesAsObject(`${dir}/${fileName}`);
+          var translations = utils.propertiesAsObject(`${dir}/${fileName}`);
 
-          return db.get(idFor(fileName))
+          return db.get(utils.idFor(fileName))
             .catch(e => {
-              if(e.status === 404) return newDocFor(fileName);
+              if(e.status === 404) return utils.newDocFor(fileName);
               else throw e;
             })
-            .then(doc => mergeProperties(doc, translations))
+            .then(doc => utils.mergeProperties(doc, translations))
             .then(doc => db.put(doc));
         }));
     });
 
 };
-
-function propertiesAsObject(path) {
-  const vals = {};
-  fs.read(path)
-    .split('\n')
-    .filter(line => line.includes('='))
-    .map(line => line.split(/=(.*)/).map(it => it.trim()).filter(it => it))
-    .map(([k, v]) => vals[k] = v);
-  return vals;
-}
-
-function mergeProperties(doc, props) {
-  if(!doc.values) doc.values = {};
-
-  for(const k in props) {
-    if(props.hasOwnProperty(k)) doc.values[k] = props[k];
-  }
-
-  return doc;
-}
-
-function newDocFor(fileName) {
-  const id = idFor(fileName);
-
-  return {
-    _id: id,
-    type: 'translations',
-    code: id.substring(id.indexOf('-') + 1),
-    name: 'TODO: please ask admin to set this in settings UI',
-    enabled: true,
-  };
-}
-
-function idFor(fileName) {
-  return fileName.substring(0, fileName.length - 11);
-}
